refactor: 도메인 메시지 레포지토의 구현체 추가
