+++ backend/ddang/src/main/java/com/ddang/ddang/chat/infrastructure/persistence/JpaMessageRepository.java
@@ -3,5 +3,5 @@ package com.ddang.ddang.chat.infrastructure.persistence;
 import com.ddang.ddang.chat.domain.Message;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface JpaMessageRepository extends JpaRepository<Message, Long>, QuerydslMessageRepository {
+public interface JpaMessageRepository extends JpaRepository<Message, Long> {
 }
